Add tests for input_buffer.c
This commit incorporates the check unit test framework into the
autotools configuration and adds a few trivial tests for
input_buffer.c. You will still be able to build the game without check
installed, but you won't be able to run the tests.

The base of the game is now built as a static library which is then
later linked to main.c. This enables us to test the main code of the
program to ensure the quality.